+++ .env.example
@@ -72,6 +72,7 @@ AUTH_SECRET=""
 # XAI_API_KEY=
 # FAL_API_KEY=
 # FIREWORKS_API_KEY=
+# LUMA_API_KEY=
 
 
 # ----------
