+++ src/ai/language/index.ts
@@ -1,0 +1,21 @@
+import {
+  openAiLanguageModels,
+  xaiLanguageModels,
+  perplexityModels,
+  googleModels,
+  anthropicModels,
+  llamaModels,
+  qwenModels,
+  deepseekModels,
+} from "./models";
+
+export const languageModels = [
+  ...anthropicModels,
+  ...googleModels,
+  ...openAiLanguageModels,
+  ...xaiLanguageModels,
+  ...perplexityModels,
+  ...llamaModels,
+  ...qwenModels,
+  ...deepseekModels,
+];
