+++ src/ai/language/models/all.ts
@@ -1,0 +1,19 @@
+import { openAiLanguageModels } from "./openai";
+import { xaiLanguageModels } from "./xai";
+import { perplexityModels } from "./perplexity";
+import { googleModels } from "./google";
+import { anthropicModels } from "./anthropic";
+import { llamaModels } from "./llama";
+import { qwenModels } from "./qwen";
+import { deepseekModels } from "./deepseek";
+
+export const allLanguageModels = [
+  ...anthropicModels,
+  ...googleModels,
+  ...openAiLanguageModels,
+  ...xaiLanguageModels,
+  ...perplexityModels,
+  ...llamaModels,
+  ...qwenModels,
+  ...deepseekModels,
+];
