+++ src/ai/language/models/deepseek.ts
@@ -1,4 +1,7 @@
-import { type LanguageModel, LanguageModelCapability } from "@/ai/types";
+import {
+  type LanguageModel,
+  LanguageModelCapability,
+} from "@/ai/language/types";
 
 const deepseekModelData: Omit<LanguageModel, "provider">[] = [
   {
