+++ src/app/_contexts/chat-context.tsx
@@ -1,6 +1,12 @@
 "use client";
 
-import { createContext, useContext, useEffect, useState, useCallback } from "react";
+import {
+  createContext,
+  useContext,
+  useEffect,
+  useState,
+  useCallback,
+} from "react";
 
 import { useChat } from "@ai-sdk/react";
 
@@ -8,29 +14,30 @@ import { toast } from "sonner";
 
 import { api } from "@/trpc/react";
 
+import { useAutoResume } from "@/app/_hooks/use-auto-resume";
+
+import { LanguageModelCapability } from "@/ai/language/types";
+
+import { clientToolkits } from "@/toolkits/toolkits/client";
+
+import { anthropicModels } from "@/ai/language/models/anthropic";
+
+import { clientCookieUtils } from "@/lib/cookies/client";
 import { generateUUID } from "@/lib/utils";
 import { fetchWithErrorHandlers } from "@/lib/fetch";
 import { ChatSDKError } from "@/lib/errors";
 
-import { useAutoResume } from "@/app/_hooks/use-auto-resume";
-
 import type { ReactNode } from "react";
 import type { Attachment, UIMessage } from "ai";
 import type { UseChatHelpers } from "@ai-sdk/react";
-import {
-  LanguageModelCapability,
-  type ImageModel,
-  type LanguageModel,
-} from "@/ai/types";
 import type { ClientToolkit } from "@/toolkits/types";
 import type { z } from "zod";
-import { clientToolkits } from "@/toolkits/toolkits/client";
 import type { SelectedToolkit } from "@/components/toolkit/types";
 import type { Toolkits } from "@/toolkits/toolkits/shared";
 import type { Workbench } from "@prisma/client";
-import { anthropicModels } from "@/ai/models/anthropic";
 import type { PersistedToolkit } from "@/lib/cookies/types";
-import { clientCookieUtils } from "@/lib/cookies/client";
+import type { ImageModel } from "@/ai/image/types";
+import type { LanguageModel } from "@/ai/language/types";
 
 const DEFAULT_CHAT_MODEL = anthropicModels[0]!;
