+++ src/app/api/chat/route.ts
@@ -34,7 +34,7 @@ import type {
 import type { Chat } from "@prisma/client";
 import { openai } from "@ai-sdk/openai";
 import { getServerToolkit } from "@/toolkits/toolkits/server";
-import { languageModels } from "@/ai/models";
+import { languageModels } from "@/ai/language";
 
 export const maxDuration = 60;
